requirements.txt
streamlit
pymupdf
python-docx
pandas
scikit-learn
sentence-transformers
plotly
jinja2
weasyprint
openai